requirements[1].txt
numpy
pandas
streamlit
pillow
scikit-learn
nltk
twython
demoji
emoji
